Maintenance-1 OOP-izat functiile
