+++ src/main/java/org/loose/fis/mov/controllers/RegistrationController.java
@@ -20,7 +20,7 @@ import java.net.URL;
 import java.util.Objects;
 import java.util.regex.Pattern;
 
-public class RegistrationController {
+public class RegistrationController extends AbstractController{
     private static final int MIN_PASSWORD_LENGTH = 8;
     @FXML
     private TextField usernameField;
@@ -94,17 +94,9 @@ public class RegistrationController {
     @FXML
     public void switchToPopUp(ActionEvent event) throws IOException {
 
-        Parent root = FXMLLoader.load(getClass().getClassLoader().getResource("changePasswordPopUp.fxml"));
-        Stage stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
-        Scene scene = new Scene(root);
-        stage.setScene(scene);
-        stage.show();
+        changeScene(event,"changePasswordPopUp.fxml");
     }
-  
-    @FXML
-    public void changePassword() {
-        System.out.println("auch!");
-    }
+
 
     private void checkMinimumPasswordStrength() throws PasswordTooWeakException {
         if (passwordField.getText().length() < MIN_PASSWORD_LENGTH) {
